Some changes to the map converter
